requirements.txt
tqdm>=4
PyYAML>=6
elasticsearch~=8.4.1
